And my version here is a 54% speedup over the original.
That means it takes less than half the time.  It achieves this by not using
this, which requires a dictionary lookup for every variable access. It uses
local variables instead, which are internally implemented in CPython as array
indexes.  Additionally, it eliminates the comparison against q by using a
range, and it removes the superfluous loop after the while.

I tried using a compact integer array, but it turns out that the math
to figure out the bit mask to use to index it is more expensive than any
savings from using the array.  I included the code anyway.