Update PrimeCPP.cpp
Sorry my system crashed yesterday up submitting this not seeing any record of it the submission today.
It should net a 15 to 20% improvement in performance. runSeive was doing several things repeatedly that didn't need to be done.
Corrected the order of initialization of variables in the class so that it gets rid of warnings at compile time.
Converted most values to uint32_t this will eliminate other warnings and conversions that aren't needed.
I would have used uint64_t except there C++ doesn't seem to handle it with %I64d which is why I dropped back to uint32_t with the exception being on line 24 the map.