Allocating Bit_Array on the heap instead of stack
For Test_Power values higher than 7 the program would crash because it
exhausted stack space. Using the heap seems to have no effect on
performance but we have much more room available to use. The program
will now run correctly for all supported powers of 10 up to 10^10!

Moved Q's calculation out of the Run_Sieve function, but it had no
effect on performance. I assume the optimizer was already doing that.

Changed the Print_Result output to more closely match the output from
other implementations. If people create docker containers to run these
tests and compare this will help them to include the Ada results since
the output is now similar.

I also changed the measuring of the Finish time to be accurate instead
of always reporting 5 seconds. 1 iteration of 10^10 was taking 82
seconds on my machine, the new results are more accurate.