Go: match bit twiddle memory handling to reduce cache pressure
Go's boolean slices are not compact bitmaps, but instead uint8 in
memory.  This means 7 of 8 bits are clogging cache without being
useful for the calculation.  Set and get using bit masking and
shifting.

Because Go memory is zeoroed on allocation, the bits are considered
"set" when they are zero.  This is the inverse of many other
implementations, due to how the Go runtime initializes memory.

Small bugfix included:  Array was only being initialized once, so the
subsequent passes were making non-op writes.  Each pass now allocates
new memory.