Implement version that runs 50% faster on my machine
I've explicitly not used constexpr shennanigans that would go against
the spirit of the competition, or used intrinsics that would pretty much
ruin the fun of the game. I've seen that there's already someone that
has provided an implementation at compile time, so I feel that the power
of C++ already showcased ;D

I also wanted to see how the alignment of the underlying data would
affect the performance. It has been a pretty cool afternoon experiment,
showing that 32-bits was the sweet spot in my machine.