Optimized Bash implementation.
Invert meaning of 0 and 1 to avoid need for slow initialization. (Uninitialized values act as 0 for the purposes of this script.)
Reduce complexity of get and clear functions for significant performance improvement.
Use associative arrays for bitArray as they're actually faster in Bash (possibly due to avoiding the expression evaluator on the index)
Upgrade to Bash 5.1 (Ubuntu 21.04) for improved performance. Note, Bash 5.1 in Alpine is about half the speed.
Use a different method of limiting the runtime of the script which doesn't require spawning a process each loop.
Removed the need for a subshell for the sqrt call.
Set LC_ALL to C to avoid localization costs.
Fixed typo on printing of times (9 digits instead of 8).
Added solution 2 which does the same inlining as PowerShell's solution 2 (albeit Bash's function call overhead isn't anywhere near as bad)