Update PrimeSieveJulia.jl
I profiled the code using @profiler and found that all of the computational cost was found in the for loop at line 37-39. 
I have created a vectorised version of that line using a range (similar to that seen in Julia Solution 2 for this project) that increases the speed of your code by ≈30 times!

Currently, on my machine, the implementation can do ≈ 200 passes. Conversely, vectorising the for loop enables it to do ≈ 6800 passes.

For convenience in demonstrating that this range is iterating through identical values as the previous for loop I have supplied a function called rangeTest, that demonstrates that the two approaches are equivalent. 

This is my first [proposed] contribution to a community Github project, which is exciting for me! I hope you like the change.

Joel