Update Dockerfile
Changing compilation from "quick optimizations" (-O2) to "slow optimizations" (-O3) to gain a bit of performance

Update PrimePas.pas

implementating some optimizations, to save few CPU cycles, resulting in a 7% performance boots.
- using 0 instead of 1 to identify a number is prime or not -> this saves the use of a not operator in the ClearBits method when calculating the mask
- removed the uneeded num variable in the RunSieve method

Update PrimePas.pas

using bit set to 1 for prime number on linux machine and bit set to 0 on Windows. For some reason on the Raspbery PI 4 the logical "and" operator seems to be faster than the "or" operator on aarch64.

Update README.md

Updating results obtained with this more optimized implementation