Haskell solution_2 add loop unrolling by block...
Increases speed about twenty percent by unrolling most of the marking loops by four marking operations, and also running the eight composite bit marking operation loops by CPU L1 cache sizes in the sieve buffer at a time for better cache associativity (cache "thrashing" is reduced although not eliminated for CPU L1 cache sizes less than 64 Kilobytes).

The `faithful to base` algorithm is still preserved as it still alternately searches for the next base prime value to be used to determine composite number multiples with actually marking the composite representation bits that are multiples of that base prime value in the sieve buffer, just that within the latter operation, the marking is done out of order for better loop and cache efficiency.  This is similar although not the same as the Rust "striping_block" implementation, with is categorized as `faithful to base` algorithm.