Add dense base primes to Chapel solution_1...
This adds a dense base prime value technique which handles base prime values up to and including nineteen; base prime values above that point use the same extreme loop unrolling technique as before.

It speeds up the extreme loop unrolling in two ways:  it reduces "cache thrashing" due to accessing memory less frequently and by larger spans closer to the cache line size, and it reduces a large proportion of the dense culling/marking operations to register/immediate value operations not involving memory access for a large reduction in the average  number of CPU clock cycles per operation.

I believe that it is `faithful to base` in that, unlike some "word" techniques it does not combine the culling/marking of composite bits within the same word with a single operation using a multi value mask but has a marking operation for each and every bit; it also assumes no prior knowledge of the base prime values other than as allowed that the only even prime is two, and covers the cases of culling by all odd step values such as nine and fifteen, which will never by used as they are not prime.

The code is long because Chapel does not have code generating macros (only something that can be used as a template) so the required dense culling patterns had to be externally code generated.  However, as Chapel is an Object Oriented Programming language, the code should be easy to read by most programmers and easily translated to other languages.

The code is issued as a program that authors of other languages can emulate the techniques used and thereby see how well optimized the code as generated by their language is; it is more interesting than the general "drag race" because authors would implement this very same technique and thus compare the compiler's ability.  Also, authors should try to write as efficiently and elegantly as possible in their chosen language, as the code will be reviewed at end culmination of this race.