Change Nim solution_3 to unrolled hybrid...
This changes this contribution to use both the old technique of "Extreme Loop Unrolling" for the sparse prime factors and the new technique of "Dense Marking" for base prime factor values below a set threshold; currently the threshold is set to odd primes below 64 for this Nim version.

This is similar to the latest Chapel hybrid version except that it uses Nim's AST macros to extend the capability to larger factors without having to use external code generation.  However, it is fast for the same reasons, it copies the sieve buffer into a register value `v`, then sets the composite multiple bits one by one with static register operations until all bits have been marked in the word for this factor, then commits the 64-bit word to the deliverable sieve buffer, repeating 64-bit words until all words in the modulo pattern are complete, then looping for all patterns across the full sieve buffer.  It is faster than the Chapel version because it has a higher "dense factor" threshold and also likely due to better compiler optimizations.