Made it easier to view macro expansions...
Added a conditional compilation to the "primes.cr" source file to short-circuit the direct view of the use of the two macros for the sparse unrolled marking and dense marking cases.  In doing so, found a minor bug that would slightly affect performance for the "extreme hybrid" technique and fixed that as well.